Extract tests for sock_*() functions into a separate file (GH-9761)
